requirements.txt
anthropic
python-dotenv
requests
tqdm
